Fixed build and cleaned up component
